Add support for app_version property
Prior to this change, Determinator didn't allow to do semver comparison on constraints. Without this it was hard to target a certain feature to a certain set of apps.

This change adds support for a new kind of constraint, app_version. An app_version constraint can be set this way:
app_version => [">=1.2.3", "<=1.3"]

The property passed to determinator would be with the following format:
app_version => ["1.2.4"]